register: prompt for confirmation
Story: As an administrator, I want to see a summary of the IPA
domain info (realm name, DNS domains, etc) before registering the
deployment with HCC, so that I can confirm I am registering the
correct IPA deployment (or stop if I am not).

Print a summary of the domain info and prompt for confirmation
before proceeding with the registration.  Add the `--unattended`
option which will skip the info message and confirmation prompt.